refactor: rename bootstrap to auth
