+++ cognite_toolkit/_cdf_tk/auth.py
